fix(api): GET /api/v1/environment new format
