+++ packages/types/lib/environment/api/index.ts
@@ -1,6 +1,7 @@
-import type { Merge } from 'type-fest';
 import type { ApiTimestamps, Endpoint } from '../../api';
 import type { DBEnvironment, DBExternalWebhook } from '../db';
+import type { ApiEnvironmentVariable } from '../variable/api';
+import type { Merge } from 'type-fest';
 
 export type ApiEnvironment = Omit<
     Merge<DBEnvironment, { callback_url: string } & ApiTimestamps>,
@@ -14,6 +15,22 @@ export type PostEnvironment = Endpoint<{
     Body: { name: string };
     Success: {
         data: Pick<DBEnvironment, 'id' | 'name'>;
+    };
+}>;
+
+export type GetEnvironment = Endpoint<{
+    Method: 'GET';
+    Path: '/api/v1/environments/current';
+    Success: {
+        environmentAndAccount: {
+            environment: ApiEnvironment;
+            env_variables: ApiEnvironmentVariable[];
+            webhook_settings: ApiWebhooks;
+            uuid: string;
+            name: string;
+            email: string;
+            slack_notifications_channel: string | null;
+        };
     };
 }>;
